fix: get xnat2bids and tests back working
